// src/app/models/repository.model.ts
export type TRepository = {
  archived: boolean;
  created_at: string;
  description: string;
  forks_count: number;
  html_url: string;
  language: string;
  name: string;
  open_issues_count: number;
  pushed_at: string;
  stargazers_count: number;
  topics: string[];
  updated_at: string;
  watchers_count: number;
};